\name{FGGMTools-package}
\alias{FGGMTools-package}
\alias{FGGMTools}
\docType{package}
\title{
\packageTitle{FGGMTools}
}
\description{
\packageDescription{FGGMTools}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{FGGMTools}
\packageIndices{FGGMTools}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{FGGMTools}

Maintainer: \packageMaintainer{FGGMTools}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
