% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IPS.R
\name{IPS}
\alias{IPS}
\title{Get a graph-constrained MLE of the covariance matrix under Gaussian Graphical Model}
\usage{
IPS(S, A, eps = 1e-08)
}
\arguments{
\item{S}{The sample covariance matrix (p by p), where p is the number of variables.}

\item{A}{An adjacency matrix corresponding to the underlyning graphical model (p by p). Must come from an undirected graph.}

\item{eps}{Error tolerance threshold (in likelihood) for the algorithm to converge}
}
\value{
\code{IPS} returns the estimated precision matrix under the graphical model.
}
\description{
Maximum likelihood estimate of the covariance matrix under the constraint of a Gaussian Graphical Model using Iterative Proportional Scaling algorithm.
}
\details{
This code uses slightly modified functions from the testing version of copCS package available at: \url{https://github.com/donaldmusgrove/copCS/blob/master/R/ips.R} by Dr. D.R. Musgrove.
}
\examples{
man/examples/IPS_ex.R
}
\references{
. Dempster AP, Covariance selection - Biometrics, 1972 - JSTOR

DR Musgrove, J Hughes, LE Eberly, Hierarchical copula regression models for areal data- Spatial Statistics, 2016 - Elsevier

Ping-Feng Xu, Jianhua Guo and Man-Lai Tang, A Localized Implementation of the Iterative Proportional Scaling Procedure for Gaussian Graphical Models. Journal of Computational and Graphical Statistics.Vol. 24, No. 1 (MARCH 2015), pp. 205-229.

Dey D., Banerjee S., Lindquist M., and Datta A., Graph-constrained Analysis for Multivariate Functional Data. Available at arXiv.org
}
\author{
Debangan Dey, Sudipto Banerjee, Martin Lindquist and Abhirup Datta
}
