% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfpca_covsel.R
\name{pfpca_covsel}
\alias{pfpca_covsel}
\title{Graph-constrained covariance estimation of multivariate function data}
\usage{
pfpca_covsel(y, t = seq(0, 1, length.out = dim(y[[1]])[2]), FVE, A)
}
\arguments{
\item{y}{list of length p containing densely observed multivariate (p-dimensional) functional data . \code{y[[j]]} is an nxm matrix of functional data for n subjects observed on a grid of length m}

\item{t}{(optional) grid on which functional data is observed, defaults to seq(0, 1, m) where \code{m = dim(data[[1]])[2]}}

\item{FVE}{Desired proportion of variance explained in the functional principal components analysis}

\item{A}{An adjacency matrix corresponding to the underlyning graphical model (p by p). Must come from an undirected graph.}
}
\value{
A list of length L (number of princtipal components explaining desired FVE), each with two variables:
\describe{
\item{\code{phi}}{A row denotes the value of l-th basis function evaluated at a grid of length m}
\item{\code{Sigma_l}}{Estimated covariance matrix for the l-th basis coefficients under graphical constraint}
}
}
\description{
Estimates the graph-constrained Karhunen-Loeve expansion for a partially separable multivariate Gaussian process.
}
\details{
This code uses slightly modified functions \code{pfpca} and some helper functions from the fgm package available at: \url{https://github.com/javzapata/fgm} by Zapata, J., Oh, S., and Petersen, A.
}
\note{
From the above list, the endowed covariance function can be obtained as \eqn{\sum_{l=1}^{L} \Sigma_l \phi_l \phi_l^T}.
}
\examples{
man/examples/pfpca_covsel_ex.R
}
\references{
Dey D., Banerjee S., Lindquist M., and Datta A., Graph-constrained Analysis for Multivariate Functional Data. Available at arXiv.org

Zapata, J., Oh, S., and Petersen, A. (2019) Functional Graphical Models for Partially Separable Gaussian Processes. Available at arXiv.org
}
\author{
Debangan Dey, Sudipto Banerjee, Martin Lindquist and Abhirup Datta
}
